start to finalize data models

+++ 2024/R/rsch_data_cseries_sept.R
@@ -34,46 +34,51 @@ base_mdl <- "2019.1b-2023" # 2023 accepted model
 # includes updated GAP tables and length-weight relationship
 base_mdl_update <- "2019.1b-2024"
 
-## copy ss input files ----
-if(!file.exists(here::here(asmnt_yr, 'rsch', base_mdl_update, 'ss3.exe'))){
-  start_ss_fldr(from = here::here(asmnt_yr - 1, 'mgmt', base_mdl),
-                to = here::here(asmnt_yr, 'rsch', base_mdl_update))
+# only run model if it doesn't exist
+if(!file.exists(here::here(asmnt_yr, 'rsch', base_mdl_update))){
+  ## copy ss input files ----
+  if(!file.exists(here::here(asmnt_yr, 'rsch', base_mdl_update, 'ss3.exe'))){
+    start_ss_fldr(from = here::here(asmnt_yr - 1, 'mgmt', base_mdl),
+                  to = here::here(asmnt_yr, 'rsch', base_mdl_update))
+  }
+  
+  # set up forecast file (with generic blocks for years so won't need to update in the future)
+  forecast <- r4ss::SS_readforecast(file = here::here(asmnt_yr, 'rsch', base_mdl_update, 'forecast.ss'))
+  forecast$Bmark_years <- c(-999, -2, -999, -2, -999, -1, -999, -2, -999, -2)
+  forecast$Fcast_years <- c(2000, -2, -5, -1, -999, -2)
+  r4ss::SS_writeforecast(mylist = forecast,
+                         dir = here::here(asmnt_yr, 'rsch', base_mdl_update),
+                         overwrite = TRUE)
+  
+  ## update files ----
+  update_ss3_files(asmnt_yr, 
+                   folder = 'rsch',
+                   mdl = base_mdl_update, 
+                   dat_filename = paste0("GOAPcod2024", dat_day, "_old.dat"),
+                   ctl_in = "Model19_1b.ctl",
+                   ctl_out = "Model19_1b.ctl")
+  
+  ## run model ----
+  run_ss3_model(asmnt_yr, 
+                folder = 'rsch',
+                mdl = base_mdl_update,
+                ctl_filename = "Model19_1b.ctl")
+  
+  
+  ## get and plot model output ----
+  # get output
+  update_base_res <- r4ss::SS_output(dir = here::here(asmnt_yr, 'rsch', base_mdl_update))
+  # if exists, delete plot folder
+  if(file.exists(here::here(asmnt_yr, 'rsch', base_mdl_update, 'plots'))){
+    unlink(here::here(asmnt_yr, 'rsch', base_mdl_update, 'plots'), recursive = TRUE)
+  }
+  # plot results
+  r4ss::SS_plots(update_base_res,
+                 printfolder = "",
+                 dir = here::here(asmnt_yr, 'rsch', base_mdl_update, "plots"))
+} else{
+  update_base_res <- r4ss::SS_output(dir = here::here(asmnt_yr, 'rsch', base_mdl_update))
 }
-
-# set up forecast file (with generic blocks for years so won't need to update in the future)
-forecast <- r4ss::SS_readforecast(file = here::here(asmnt_yr, 'rsch', base_mdl_update, 'forecast.ss'))
-forecast$Bmark_years <- c(-999, -2, -999, -2, -999, -1, -999, -2, -999, -2)
-forecast$Fcast_years <- c(2000, -2, -5, -1, -999, -2)
-r4ss::SS_writeforecast(mylist = forecast,
-                       dir = here::here(asmnt_yr, 'rsch', base_mdl_update),
-                       overwrite = TRUE)
-
-## update files ----
-update_ss3_files(asmnt_yr, 
-                 folder = 'rsch',
-                 mdl = base_mdl_update, 
-                 dat_filename = paste0("GOAPcod2024", dat_day, "_old.dat"),
-                 ctl_in = "Model19_1b.ctl",
-                 ctl_out = "Model19_1b.ctl")
-
-## run model ----
-run_ss3_model(asmnt_yr, 
-              folder = 'rsch',
-              mdl = base_mdl_update,
-              ctl_filename = "Model19_1b.ctl")
-
-
-## get and plot model output ----
-# get output
-update_base_res <- r4ss::SS_output(dir = here::here(asmnt_yr, 'rsch', base_mdl_update))
-# if exists, delete plot folder
-if(file.exists(here::here(asmnt_yr, 'rsch', base_mdl_update, 'plots'))){
-  unlink(here::here(asmnt_yr, 'rsch', base_mdl_update, 'plots'), recursive = TRUE)
-}
-# plot results
-r4ss::SS_plots(update_base_res,
-               printfolder = "",
-               dir = here::here(asmnt_yr, 'rsch', base_mdl_update, "plots"))
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # 2019.1c.1 ----
@@ -531,9 +536,9 @@ r4ss::SS_plots(new_base.8_res,
 new_base <- "2019.1c-2024"
 
 ## copy ss input files ----
-if(!file.exists(here::here(asmnt_yr, 'rsch', 'cseries', new_base, 'ss3.exe'))){
+if(!file.exists(here::here(asmnt_yr, 'rsch', new_base, 'ss3.exe'))){
   start_ss_fldr(from = here::here(asmnt_yr, 'rsch', base_mdl_update),
-                to = here::here(asmnt_yr, 'rsch', 'cseries', new_base))
+                to = here::here(asmnt_yr, 'rsch', new_base))
 }
 
 ## update files ----
@@ -558,15 +563,15 @@ run_ss3_model(asmnt_yr,
 
 ## get and plot model output ----
 # get output
-new_base_res <- r4ss::SS_output(dir = here::here(asmnt_yr, 'rsch', 'cseries', new_base))
+new_base_res <- r4ss::SS_output(dir = here::here(asmnt_yr, 'rsch', new_base))
 # if exists, delete plot folder
-if(file.exists(here::here(asmnt_yr, 'rsch', 'cseries', new_base, 'plots'))){
-  unlink(here::here(asmnt_yr, 'rsch', 'cseries', new_base, 'plots'), recursive = TRUE)
+if(file.exists(here::here(asmnt_yr, 'rsch', new_base, 'plots'))){
+  unlink(here::here(asmnt_yr, 'rsch', new_base, 'plots'), recursive = TRUE)
 }
 # plot results
 r4ss::SS_plots(new_base_res,
                printfolder = "",
-               dir = here::here(asmnt_yr, 'rsch', 'cseries', new_base, "plots"))
+               dir = here::here(asmnt_yr, 'rsch', new_base, "plots"))
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # plot comparisons ----
